+++ src/libcollections/lib.rs
@@ -46,7 +46,6 @@ pub use alloc::binary_heap;
 pub use alloc::borrow;
 pub use alloc::fmt;
 pub use alloc::linked_list;
-#[allow(deprecated)]
 pub use alloc::range;
 pub use alloc::slice;
 pub use alloc::str;
