+++ src/libcore/ops/mod.rs
@@ -183,9 +183,6 @@ pub use self::index::{Index, IndexMut};
 #[stable(feature = "rust1", since = "1.0.0")]
 pub use self::range::{Range, RangeFrom, RangeFull, RangeTo};
 
-#[stable(feature = "range_argument", since = "1.19.0")]
-pub use self::range::{RangeArgument, Bound};
-
 #[unstable(feature = "inclusive_range", reason = "recently added, follows RFC", issue = "28237")]
 pub use self::range::{RangeInclusive, RangeToInclusive};
