+++ src/librustc_data_structures/accumulate_vec.rs
@@ -15,10 +15,11 @@
 //!
 //! The N above is determined by Array's implementor, by way of an associatated constant.
 
-use std::ops::{Deref, DerefMut, RangeArgument};
+use std::ops::{Deref, DerefMut};
 use std::iter::{self, IntoIterator, FromIterator};
 use std::slice;
 use std::vec;
+use std::collections::range::RangeArgument;
 
 use rustc_serialize::{Encodable, Encoder, Decodable, Decoder};
