+++ src/librustc_data_structures/indexed_vec.rs
@@ -8,11 +8,12 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+use std::collections::range::RangeArgument;
 use std::fmt::Debug;
 use std::iter::{self, FromIterator};
 use std::slice;
 use std::marker::PhantomData;
-use std::ops::{Index, IndexMut, Range, RangeArgument};
+use std::ops::{Index, IndexMut, Range};
 use std::fmt;
 use std::vec;
 use std::u32;
