+++ src/librustc_data_structures/lib.rs
@@ -25,6 +25,7 @@
 #![deny(warnings)]
 
 #![feature(shared)]
+#![feature(collections_range)]
 #![feature(nonzero)]
 #![feature(unboxed_closures)]
 #![feature(fn_traits)]
